add token-auth impl to upload-del albums

+++ src/components/settings/Settings.jsx
@@ -2,6 +2,7 @@
 
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import { auth } from '../utils/Firebase';
 import { Input } from '../../common/input';
 import { Button } from '../../common/button';
 import SnackAlert from '../../common/SnackAlert';
@@ -33,13 +34,20 @@ const Settings = () => {
 
   const deleteAlbums = async () => {
     try {
-      const del_response = await axios.delete(`${BASE_API_URI}/delete-albums`, {
-        params: {
-          user_email: email,
-        },
-      });
-      setSnackMessage(del_response.data?.message);
-      setSnackState('success');
+      const user = auth.currentUser;
+      if (user) {
+        const idToken = await user.getIdToken();
+        const del_response = await axios.delete(`${BASE_API_URI}/delete-albums`, {
+          params: {
+            user_email: email,
+          },
+          headers: {
+            Authorization: `Bearer ${idToken}`,
+          },
+        });
+        setSnackMessage(del_response.data?.message);
+        setSnackState('success');
+      }
     } catch (err) {
       setSnackMessage(err.message);
       setSnackState('error');
